Rename object map to reflect the fact that it is merely an entrypoint
